+++ packages/Webkul/Admin/src/DataGrids/Marketing/SitemapDataGrid.php
@@ -75,23 +75,27 @@ class SitemapDataGrid extends DataGrid
      */
     public function prepareActions()
     {
-        $this->addAction([
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.marketing.sitemaps.index.datagrid.edit'),
-            'method' => 'GET',
-            'route'  => 'admin.marketing.promotions.sitemaps.update',
-            'url'    => function ($row) {
-                return route('admin.marketing.promotions.sitemaps.update', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('marketing.sitemaps.edit')) {
+            $this->addAction([
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.marketing.sitemaps.index.datagrid.edit'),
+                'method' => 'GET',
+                'route'  => 'admin.marketing.promotions.sitemaps.update',
+                'url'    => function ($row) {
+                    return route('admin.marketing.promotions.sitemaps.update', $row->id);
+                },
+            ]);
+        }
 
-        $this->addAction([
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.marketing.sitemaps.index.datagrid.delete'),
-            'method' => 'DELETE',
-            'url'    => function ($row) {
-                return route('admin.marketing.promotions.sitemaps.delete', $row->id);
-            },
-        ]);
+        if (bouncer()->hasPermission('marketing.sitemaps.delete')) {
+            $this->addAction([
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.marketing.sitemaps.index.datagrid.delete'),
+                'method' => 'DELETE',
+                'url'    => function ($row) {
+                    return route('admin.marketing.promotions.sitemaps.delete', $row->id);
+                },
+            ]);
+        }
     }
 }
