+++ packages/Webkul/Admin/src/Resources/views/marketing/sitemaps/index.blade.php
@@ -12,9 +12,11 @@
             </p>
 
             <!-- Create Button -->
-            <div class="primary-button">
-                @lang('admin::app.marketing.sitemaps.index.create-btn')
-            </div>
+            @if (bouncer()->hasPermission('marketing.sitemaps.create'))
+                <div class="primary-button">
+                    @lang('admin::app.marketing.sitemaps.index.create-btn')
+                </div>
+            @endif
         </div>
 
         {{-- Added For Shimmer --}}
@@ -32,21 +34,27 @@
                 </p>
 
                 <!-- Create Button -->
-                <div 
-                    class="primary-button"
-                    @click="selectedSitemap=0; $refs.sitemap.toggle()"
-                >
-                    @lang('admin::app.marketing.sitemaps.index.create-btn')
-                </div>
+                @if (bouncer()->hasPermission('marketing.sitemaps.create'))
+                    <div 
+                        class="primary-button"
+                        @click="selectedSitemap=0; $refs.sitemap.toggle()"
+                    >
+                        @lang('admin::app.marketing.sitemaps.index.create-btn')
+                    </div>
+                @endif
             </div>
 
             <x-admin::datagrid
                 src="{{ route('admin.marketing.promotions.sitemaps.index') }}"
                 ref="datagrid"
             >
+                @php
+                    $hasPermission = bouncer()->hasPermission('marketing.sitemaps.edit') || bouncer()->hasPermission('marketing.sitemaps.delete');
+                @endphp
+
                 <!-- Datagrid Header -->
                 <template #header="{ columns, records, sortPage, applied }">
-                    <div class="row grid grid-cols-5 grid-rows-1 gap-[10px] items-center px-[16px] py-[10px] border-b-[1px] border-gray-300 text-gray-600 bg-gray-50 font-semibold">
+                    <div class="row grid grid-cols-{{ $hasPermission ? '5' : '4' }} grid-rows-1 gap-[10px] items-center px-[16px] py-[10px] border-b-[1px] border-gray-300 text-gray-600 bg-gray-50 font-semibold">
                         <div
                             class="flex gap-[10px] cursor-pointer"
                             v-for="(columnGroup, index) in ['id', 'file_name', 'path', 'url']"
@@ -77,9 +85,11 @@
                         </div>
 
                         <!-- Actions -->
-                        <p class="col-start-[none]">
-                            @lang('admin::app.components.datagrid.table.actions')
-                        </p>
+                        @if ($hasPermission)
+                            <p class="col-start-[none]">
+                                @lang('admin::app.components.datagrid.table.actions')
+                            </p>
+                        @endif
                     </div>
                 </template>
 
@@ -88,7 +98,7 @@
                     <div
                         v-for="record in records"
                         class="row grid gap-[10px] items-center px-[16px] py-[16px] border-b-[1px] border-gray-300 text-gray-600 transition-all hover:bg-gray-50"
-                        style="grid-template-columns: repeat(5, 1fr);"
+                        :style="'grid-template-columns: repeat(' + (record.actions.length ? 5 : 4) + ', 1fr);'"
                     >
                         <!-- Id -->
                         <p v-text="record.id"></p>
@@ -108,23 +118,27 @@
 
                         <!-- Actions -->
                         <div class="flex justify-end">
-                            <a @click="selectedSitemap=1; editModal(record)">
-                                <span
-                                    :class="record.actions['0'].icon"
-                                    class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-100 max-sm:place-self-center"
-                                    :title="record.actions['0'].title"
-                                >
-                                </span>
-                            </a>
-
-                            <a @click="deleteModal(record.actions['1']?.url)">
-                                <span
-                                    :class="record.actions['1'].icon"
-                                    class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-100 max-sm:place-self-center"
-                                    :title="record.actions['1'].title"
-                                >
-                                </span>
-                            </a>
+                            @if (bouncer()->hasPermission('marketing.sitemaps.edit'))
+                                <a @click="selectedSitemap=1; editModal(record)">
+                                    <span
+                                        :class="record.actions['0'].icon"
+                                        class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-100 max-sm:place-self-center"
+                                        :title="record.actions['0'].title"
+                                    >
+                                    </span>
+                                </a>
+                            @endif
+                            
+                            @if (bouncer()->hasPermission('marketing.sitemaps.delete'))
+                                <a @click="deleteModal(record.actions['1']?.url)">
+                                    <span
+                                        :class="record.actions['1'].icon"
+                                        class="cursor-pointer rounded-[6px] p-[6px] text-[24px] transition-all hover:bg-gray-100 max-sm:place-self-center"
+                                        :title="record.actions['1'].title"
+                                    >
+                                    </span>
+                                </a>
+                            @endif
                         </div>
                     </div>
                 </template>
